refactor: extract shared heap snapshot abstraction into src/snapshot/
